cohorts/2025/04-analytics-engineering/requirements.txt
dlt
pandas
requests
google-cloud-storage
google-auth